refactor: refactor and clean the code ( server )
