+++ crypto-app/src/app/api/auth/[...nextauth]/route.ts
@@ -1,5 +1,5 @@
 import NextAuth from "next-auth"
-import { authOptions } from "@/utils/auth"
+import { authOptions } from "@/utils/auth/auth"
 
 const handler = NextAuth(authOptions)
