+++ crypto-app/src/app/crypto/favorites/page.tsx
@@ -1,9 +1,9 @@
 import { Button } from "@/components/ui/button"
 import { CryptosRealTimeTable } from "@/components/dashborad/crypto/CryptosRealTimeTable"
 import Link from "next/link"
-import { trpcServerComp } from "@/server/trpc.Provider"
+import { trpcServerComp } from "@/server/trpcProvider"
 import { getServerSession } from "next-auth"
-import { authOptions} from "@/utils/auth"
+import { authOptions} from "@/utils/auth/auth"
 import { redirect } from "next/navigation"
 import { ConvertCryptoSection } from "@/components/dashborad/crypto/ConvertCryptoSection"
