+++ crypto-app/src/components/dashborad/crypto/CryptosRealTimeTable.tsx
@@ -16,7 +16,7 @@ import {
   DialogTrigger,
 } from "@/components/ui/dialog"
 import { Input } from "@/components/ui/input";
-import { trpcClientComp } from "@/server/trpc.Provider";
+import { trpcClientComp } from "@/server/trpcProvider";
 import { toast } from "sonner";
 import { useSession } from "next-auth/react";
 import { Button } from "@/components/ui/button";
