+++ crypto-app/src/components/dashborad/currencies/AllCurrencySymbols.tsx
