+++ crypto-app/src/config/db/index.ts
@@ -1,6 +1,6 @@
 import { drizzle } from "drizzle-orm/postgres-js";
 import postgres from "postgres";
-import * as schema from "@/db/schema";
+import * as schema from "@/config/db/schema";
 import { envVars } from "@/config/envVars/envVars.config";
