+++ crypto-app/src/config/db/schema/index.ts
