+++ crypto-app/src/config/db/schema/user.ts
