+++ crypto-app/src/config/db/schema/userCrypto.ts
