+++ crypto-app/src/config/db/schema/userCurrency.ts
