+++ crypto-app/src/stores/authStore.ts
